single init for read-only integration tests
