+++ stats/stats-server/tests/it/chart_endpoints/lines.rs
@@ -1,54 +1,11 @@
-use blockscout_service_launcher::{
-    launcher::ConfigSettings,
-    test_server::{get_test_server_settings, init_server, send_get_request},
-};
-use chrono::NaiveDate;
+use blockscout_service_launcher::test_server::send_get_request;
 
-use stats::{
-    tests::{
-        init_db::init_db_all,
-        mock_blockscout::{default_mock_blockscout_api, fill_mock_blockscout_data},
-    },
-    ResolutionKind,
-};
-use stats_server::{stats, Settings};
+use stats::ResolutionKind;
+use url::Url;
 
-use std::{collections::HashMap, path::PathBuf, str::FromStr};
+use crate::common::enabled_resolutions;
 
-pub(crate) async fn enabled_resolutions(
-    line_charts: stats_proto::blockscout::stats::v1::LineCharts,
-) -> HashMap<String, Vec<String>> {
-    line_charts
-        .sections
-        .iter()
-        .flat_map(|sec| sec.charts.clone())
-        .map(|l| (l.id, l.resolutions))
-        .collect()
-}
-
-#[tokio::test]
-#[ignore = "needs database"]
-async fn test_lines_ok() {
-    let (stats_db, blockscout_db) = init_db_all("test_lines_ok").await;
-    let blockscout_api = default_mock_blockscout_api().await;
-    fill_mock_blockscout_data(&blockscout_db, NaiveDate::from_str("2023-03-01").unwrap()).await;
-
-    std::env::set_var("STATS__CONFIG", "./tests/config/test.toml");
-    let mut settings = Settings::build().expect("Failed to build settings");
-    let (server_settings, base) = get_test_server_settings();
-    settings.server = server_settings;
-    settings.charts_config = PathBuf::from_str("../config/charts.json").unwrap();
-    settings.layout_config = PathBuf::from_str("../config/layout.json").unwrap();
-    settings.update_groups_config = PathBuf::from_str("../config/update_groups.json").unwrap();
-    settings.db_url = stats_db.db_url();
-    settings.blockscout_db_url = blockscout_db.db_url();
-    settings.blockscout_api_url = Some(url::Url::from_str(&blockscout_api.uri()).unwrap());
-
-    init_server(|| stats(settings), &base).await;
-
-    // Sleep until server will start and calculate all values
-    tokio::time::sleep(std::time::Duration::from_secs(7)).await;
-
+pub async fn test_lines_ok(base: Url) {
     let line_charts: stats_proto::blockscout::stats::v1::LineCharts =
         send_get_request(&base, "/api/v1/lines").await;
